Move plugins/disco.py to _internal/
