+++ certbot/plugins/selection_test.py
@@ -11,7 +11,7 @@ from certbot import errors
 from certbot import interfaces
 from certbot.compat import os
 from certbot.display import util as display_util
-from certbot.plugins.disco import PluginsRegistry
+from certbot._internal.plugins.disco import PluginsRegistry
 from certbot.tests import util as test_util
