--- docs/api/plugins/disco.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.plugins.disco`
---------------------------------
-
-.. automodule:: certbot.plugins.disco
-   :members:
